requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Werkzeug==2.0.1
PyJWT==2.1.0
python-dotenv==0.19.0
gunicorn==20.1.0
